# internal/api/swagger/swagger.yaml
basePath: /v1
definitions:
...
  internal_api_handler.envelope:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  internal_api_handler.journalRequest:
    properties:
      content:
        type: string
      occurred_at_date:
        description: '"YYYY-MM-DD"'
        type: string
      occurred_at_time:
        description: '"HH:MM" or ""'
        type: string
      person_ids:
        items:
          type: integer
        type: array
      title:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: kith
    url: https://github.com/nhymxu/kith-pms
  description: Self-hosted personal relationship manager API.
  license:
    name: MIT
  title: kith API
  version: "1.0"
paths:
...
  /journal:
    get:
      parameters:
      - description: Full-text search query
        in: query
        name: q
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: From date YYYY-MM-DD
        in: query
        name: from_date
        type: string
      - description: To date YYYY-MM-DD
        in: query
        name: to_date
        type: string
      - description: Comma-separated person IDs
        in: query
        name: person_ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: List journal entries
      tags:
      - journal
    post:
      consumes:
      - application/json
      parameters:
      - description: Journal entry
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.journalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    id:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Create journal entry
      tags:
      - journal
  /journal/{id}:
    delete:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Delete journal entry
      tags:
      - journal
    get:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Get journal entry
      tags:
      - journal
    put:
      consumes:
      - application/json
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Journal entry
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.journalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    id:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Update journal entry
      tags:
      - journal
...
  /me:
    get:
      description: Returns the person designated as "Me", or 404 if not configured.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Get self profile
      tags:
      - me
  /me/setup:
    post:
      consumes:
      - application/json
      description: Designates an existing person as the "Me" person.
      parameters:
      - description: Person to designate as self
        in: body
        name: body
        required: true
        schema:
          properties:
            person_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    person_id:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Setup self profile
      tags:
      - me
...
securityDefinitions:
  CSRFHeader:
    description: Must be set to "kith-spa" on all mutating requests (POST/PUT/PATCH/DELETE)
    in: header
    name: X-Requested-With
    type: apiKey
  CookieAuth:
    description: HttpOnly session cookie set by POST /v1/auth/login
    in: cookie
    name: kith_session
    type: apiKey
swagger: "2.0"